fix(cardano-chain-follower): break down stats
