+++ rust/cardano-chain-follower/src/stats/mod.rs
@@ -1,180 +1,23 @@
 //! Cardano Chain Follower Statistics
 
+pub(crate) mod follower;
+pub(crate) mod live_chain;
+pub(crate) mod mithril;
+pub(crate) mod rollback;
+
 use std::sync::{Arc, LazyLock, RwLock};
 
 use cardano_blockchain_types::{Network, Slot};
-use chrono::{DateTime, Utc};
+use chrono::Utc;
 use dashmap::DashMap;
+use rollback::{rollbacks, rollbacks_reset, RollbackType};
 use serde::Serialize;
-use strum::{EnumIter, IntoEnumIterator};
+use strum::IntoEnumIterator;
 use tracing::error;
 
+use crate::stats::{live_chain::Live, mithril::Mithril};
+
 // -------- GENERAL STATISTIC TRACKING
-
-/// Statistics related to Mithril Snapshots
-#[derive(Debug, Default, Clone, Serialize)]
-pub struct Mithril {
-    /// Number of Mithril Snapshots that have downloaded successfully.
-    pub updates: u64,
-    /// The Immutable TIP Slot# - Origin = No downloaded snapshot
-    pub tip: Slot,
-    /// Time we started downloading the current snapshot. 1/1/1970-00:00:00 UTC = Never
-    /// downloaded.
-    pub dl_start: DateTime<Utc>,
-    /// Time we finished downloading the current snapshot. if < `dl_start` its the
-    /// previous time we finished.
-    pub dl_end: DateTime<Utc>,
-    /// Number of times download failed (bad server connection)
-    pub dl_failures: u64,
-    /// The time the last download took, in seconds.
-    pub last_dl_duration: u64,
-    /// The size of the download archive, in bytes. (If not started and not ended, current
-    /// partial download size).
-    pub dl_size: u64,
-    /// Extraction start time. 1/1/1970-00:00:00 UTC = Never extracted.
-    pub extract_start: DateTime<Utc>,
-    /// Extraction end time. if `extract_end` < `extract_start` its the previous time we
-    /// finished extracting.
-    pub extract_end: DateTime<Utc>,
-    /// Number of times extraction failed (bad archive)
-    pub extract_failures: u64,
-    /// Size of last extracted snapshot, in bytes.
-    pub extract_size: u64,
-    /// Deduplicated Size vs previous snapshot.
-    pub deduplicated_size: u64,
-    /// Number of identical files deduplicated from previous snapshot.
-    pub deduplicated: u64,
-    /// Number of changed files from previous snapshot.
-    pub changed: u64,
-    /// Number of new files from previous snapshot.
-    pub new: u64,
-    /// Mithril Certificate Validation Start Time. 1/1/1970-00:00:00 UTC = Never
-    /// validated.
-    pub validate_start: DateTime<Utc>,
-    /// Mithril Certificate Validation End Time. if validate end < validate start its the
-    /// previous time we finished validating.
-    pub validate_end: DateTime<Utc>,
-    /// Number of times validation failed (bad snapshot)
-    pub validate_failures: u64,
-    /// Blocks that failed to deserialize from the mithril immutable chain.
-    pub invalid_blocks: u64,
-    /// Download Or Validation Failed
-    pub download_or_validation_failed: u64,
-    /// Failed to get tip from mithril snapshot.
-    pub failed_to_get_tip: u64,
-    /// Tip failed to advance
-    pub tip_did_not_advance: u64,
-    /// Failed to send new tip to updater.
-    pub tip_failed_to_send_to_updater: u64,
-    /// Failed to activate new snapshot
-    pub failed_to_activate_new_snapshot: u64,
-}
-
-impl Mithril {
-    /// Reset incremental counters in the mithril statistics.
-    fn reset(&mut self) {
-        self.updates = 0;
-        self.dl_failures = 0;
-        self.extract_failures = 0;
-        self.validate_failures = 0;
-        self.invalid_blocks = 0;
-    }
-}
-
-/// Statistics related to a single depth of rollback
-#[derive(Debug, Default, Clone, Serialize)]
-pub struct Rollback {
-    /// How deep was the rollback from tip.
-    pub depth: u64,
-    /// How many times has a rollback been this deep.
-    pub count: u64,
-}
-
-/// Statistics for all our known rollback types
-/// Rollback Vec is sorted by depth, ascending.
-#[derive(Debug, Default, Clone, Serialize)]
-pub struct Rollbacks {
-    /// These are the ACTUAL rollbacks we did on our live-chain in memory.
-    pub live: Vec<Rollback>,
-    /// These are the rollbacks reported by the Peer Node, which may not == an actual
-    /// rollback on our internal live chain.
-    pub peer: Vec<Rollback>,
-    /// These are the rollbacks synthesized for followers, based on their reading of the
-    /// chain tip.
-    pub follower: Vec<Rollback>,
-}
-
-/// Individual Follower stats
-#[derive(Debug, Default, Clone, Serialize)]
-pub struct Follower {
-    /// Synthetic follower connection ID
-    pub id: u64,
-    /// Starting slot for this follower (0 = Start at Genesis Block for the chain).
-    pub start: Slot,
-    /// Current slot for this follower.
-    pub current: Slot,
-    /// Target slot for this follower (MAX U64 == Follow Tip Forever).
-    pub end: Slot,
-    /// Current Sync Time.
-    pub sync_start: DateTime<Utc>,
-    /// When this follower reached TIP or its destination slot.
-    pub sync_end: Option<DateTime<Utc>>,
-}
-
-/// Statistics related to the live blockchain
-#[derive(Debug, Default, Clone, Serialize)]
-pub struct Live {
-    /// The Time that synchronization to this blockchain started
-    pub sync_start: DateTime<Utc>,
-    /// The Time that synchronization to this blockchain was complete up-to-tip. None =
-    /// Not yet synchronized.
-    pub sync_end: Option<DateTime<Utc>>,
-    /// When backfill started
-    pub backfill_start: Option<DateTime<Utc>>,
-    /// Backfill size to achieve synchronization. (0 before sync completed)
-    pub backfill_size: u64,
-    /// When backfill ended
-    pub backfill_end: Option<DateTime<Utc>>,
-    /// Backfill Failures
-    pub backfill_failures: u64,
-    /// The time of the last backfill failure
-    pub backfill_failure_time: Option<DateTime<Utc>>,
-    /// Current Number of Live Blocks
-    pub blocks: u64,
-    /// The current head of the live chain slot#
-    pub head_slot: Slot,
-    /// The current live tip slot# as reported by the peer.
-    pub tip: Slot,
-    /// Number of times we connected/re-connected to the Node.
-    pub reconnects: u64,
-    /// Last reconnect time,
-    pub last_connect: DateTime<Utc>,
-    /// Last reconnect time,
-    pub last_connected_peer: String,
-    /// Last disconnect time,
-    pub last_disconnect: DateTime<Utc>,
-    /// Last disconnect time,
-    pub last_disconnected_peer: String,
-    /// Is there an active connection to the node
-    pub connected: bool,
-    /// Rollback statistics.
-    pub rollbacks: Rollbacks,
-    /// New blocks read from blockchain.
-    pub new_blocks: u64,
-    /// Blocks that failed to deserialize from the blockchain.
-    pub invalid_blocks: u64,
-    /// Active Followers (range and current depth)
-    pub follower: Vec<Follower>,
-}
-
-impl Live {
-    /// Reset incremental counters in the live statistics.
-    fn reset(&mut self) {
-        self.new_blocks = 0;
-        self.reconnects = 0;
-        self.invalid_blocks = 0;
-    }
-}
 
 /// Statistics for a single follower network.
 #[derive(Debug, Default, Clone, Serialize)]
@@ -613,116 +456,6 @@ pub(crate) fn mithril_sync_failure(network: Network, failure: MithrilSyncFailure
             chain_stats.mithril.failed_to_activate_new_snapshot += 1;
         },
     }
-}
-
-// -------- ROLLBACK STATISTIC TRACKING
-// ----------------------------------------------------------
-
-/// The types of rollbacks we track for a chain.
-#[derive(EnumIter, Eq, Ord, PartialEq, PartialOrd, Copy, Clone, Hash)]
-pub enum RollbackType {
-    /// Rollback on the in-memory live chain.
-    LiveChain,
-    /// Rollback signaled by the peer.
-    Peer,
-    /// Rollback synthesized for the Follower.
-    Follower,
-}
-
-/// Individual rollback records.
-type RollbackRecords = DashMap<u64, Rollback>;
-/// Rollback Records per rollback type.
-type RollbackTypeMap = DashMap<RollbackType, Arc<RwLock<RollbackRecords>>>;
-/// Record of rollbacks.
-type RollbackMap = DashMap<Network, RollbackTypeMap>;
-/// Statistics of rollbacks detected per chain.
-static ROLLBACKS_MAP: LazyLock<RollbackMap> = LazyLock::new(|| {
-    let map = RollbackMap::new();
-    for network in Network::iter() {
-        let type_map = RollbackTypeMap::new();
-        for rollback in RollbackType::iter() {
-            type_map.insert(rollback, Arc::new(RwLock::new(RollbackRecords::new())));
-        }
-        map.insert(network, type_map);
-    }
-    map
-});
-
-/// Get the actual rollback map for a chain.
-fn lookup_rollback_map(
-    network: Network, rollback: RollbackType,
-) -> Option<Arc<RwLock<RollbackRecords>>> {
-    let Some(chain_rollback_map) = ROLLBACKS_MAP.get(&network) else {
-        error!("Rollback stats SHOULD BE exhaustively pre-allocated.");
-        return None;
-    };
-    let chain_rollback_map = chain_rollback_map.value();
-
-    let Some(rollback_map) = chain_rollback_map.get(&rollback) else {
-        error!("Rollback stats SHOULD BE exhaustively pre-allocated.");
-        return None;
-    };
-    let rollback_map = rollback_map.value();
-
-    Some(rollback_map.clone())
-}
-
-/// Extract the current rollback stats as a vec.
-fn rollbacks(network: Network, rollback: RollbackType) -> Vec<Rollback> {
-    let Some(rollback_map) = lookup_rollback_map(network, rollback) else {
-        return Vec::new();
-    };
-
-    let Ok(rollback_values) = rollback_map.read() else {
-        error!("Rollback stats LOCK Poisoned, should not happen.");
-        return vec![];
-    };
-
-    let mut rollbacks = Vec::new();
-
-    // Get all the rollback stats.
-    for stat in rollback_values.iter() {
-        rollbacks.push(stat.value().clone());
-    }
-
-    rollbacks
-}
-
-/// Reset ALL the rollback stats for a given blockchain.
-fn rollbacks_reset(network: Network, rollback: RollbackType) -> Vec<Rollback> {
-    let Some(rollback_map) = lookup_rollback_map(network, rollback) else {
-        return Vec::new();
-    };
-
-    let Ok(rollbacks) = rollback_map.write() else {
-        error!("Rollback stats LOCK Poisoned, should not happen.");
-        return vec![];
-    };
-
-    rollbacks.clear();
-
-    Vec::new()
-}
-
-/// Count a rollback
-pub(crate) fn rollback(network: Network, rollback: RollbackType, depth: u64) {
-    let Some(rollback_map) = lookup_rollback_map(network, rollback) else {
-        return;
-    };
-
-    let Ok(rollbacks) = rollback_map.write() else {
-        error!("Rollback stats LOCK Poisoned, should not happen.");
-        return;
-    };
-
-    let mut value = match rollbacks.get(&depth) {
-        Some(value_entry) => (*value_entry.value()).clone(),
-        None => Rollback { depth, count: 0 },
-    };
-
-    value.count += 1;
-
-    let _unused = rollbacks.insert(depth, value);
 }
 
 #[cfg(test)]
